+++ app/src/main/java/com/example/filehunt/DocsActivityRe.java
@@ -38,8 +38,10 @@ import com.example.filehunt.Model.Model_Docs;
 import com.example.filehunt.Utils.AlertDialogHelper;
 import com.example.filehunt.Utils.AsynctaskUtility;
 import com.example.filehunt.Utils.AutoFitGridLayoutManager;
+import com.example.filehunt.Utils.CustomProgressDialog;
 import com.example.filehunt.Utils.RecyclerItemClickListener;
 import com.example.filehunt.Utils.Utility;
+import com.example.filehunt.Utils.UtilityStorage;
 
 import java.io.File;
 import java.util.ArrayList;
@@ -79,6 +81,7 @@ public class DocsActivityRe extends AppCompatActivity implements AlertDialogHelp
         blankIndicator=(ImageView) findViewById(R.id.blankIndicator);
         mcontext=DocsActivityRe.this;
         instance=this;
+        UtilityStorage.InitilaizePrefs(mcontext);
         Utility.setActivityTitle(mcontext,getResources().getString(R.string.cat_Documents));
        
 //           int_position = getIntent().getIntExtra("value", 0);
@@ -124,6 +127,33 @@ public class DocsActivityRe extends AppCompatActivity implements AlertDialogHelp
 
 
     }
+
+    @Override
+    protected void onActivityResult(int requestCode, int resultCode, Intent data) {
+        super.onActivityResult(requestCode, resultCode, data);
+
+        if(requestCode==Constants.FILE_DELETE_REQUEST_CODE) {
+            boolean isPersistUriSet = UtilityStorage.setUriForStorage(requestCode, resultCode, data);
+            if (isPersistUriSet && multiselect_list.size() > 0)
+                new DeleteFileTask(multiselect_list).execute();
+        }
+        if(requestCode==Constants.FILE_RENAME_REQUEST_CODE)
+        {
+            boolean isPersistUriSet = UtilityStorage.setUriForStorage(requestCode, resultCode, data);
+            if (isPersistUriSet && multiselect_list.size() > 0)
+            {
+                // call rename function  here in the case of premission granted first  time;
+
+                Utility.renameFile(mcontext,multiselect_list.get(0).getFilePath(),Constants.Global_File_Rename_NewName,3);
+            }
+
+        }
+
+
+
+
+    }
+
     public static DocsActivityRe getInstance() {
         return instance;
     }
@@ -358,15 +388,33 @@ public class DocsActivityRe extends AppCompatActivity implements AlertDialogHelp
         {
             if(multiselect_list.size()>0)
             {
-                new  DeleteFileTask(multiselect_list).execute();
-                for(int i=0;i<multiselect_list.size();i++)
-                    docsList.remove(multiselect_list.get(i));
 
-                multiSelectAdapter.notifyDataSetChanged();
 
-                if (mActionMode != null) {
-                    mActionMode.finish();
+                //
+                File f =new File(multiselect_list.get(0).getFilePath());
+
+                if(UtilityStorage.isWritableNormalOrSaf(f,mcontext)) {
+                    new DeleteFileTask(multiselect_list).execute();
                 }
+                else
+                {
+                    UtilityStorage.guideDialogForLEXA(mcontext,f.getParent(),Constants.FILE_DELETE_REQUEST_CODE);
+                }
+
+
+
+                // // now this task  is being done  on post execute of delete task
+
+//                for(int i=0;i<multiselect_list.size();i++)
+//                    docsList.remove(multiselect_list.get(i));
+//
+//                multiSelectAdapter.notifyDataSetChanged();
+//
+//                if (mActionMode != null) {
+//                    mActionMode.finish();
+//                }
+
+
 
             }
         }
@@ -420,6 +468,12 @@ public class DocsActivityRe extends AppCompatActivity implements AlertDialogHelp
 
     private class DeleteFileTask extends AsyncTask<Void,Void,Integer>
     {
+        @Override
+        protected void onPreExecute() {
+            super.onPreExecute();
+            CustomProgressDialog.show(mcontext,getResources().getString(R.string.deleting_file));
+        }
+
         ArrayList<Model_Docs> multiselect_list;
         DeleteFileTask( ArrayList<Model_Docs> multiselect_list)
         {
@@ -436,7 +490,22 @@ public class DocsActivityRe extends AppCompatActivity implements AlertDialogHelp
         protected void onPostExecute(Integer FileCount) {
             super.onPostExecute(FileCount);
 
+
+            // remove  the  file from docsList list if deleted;
+
+            if(FileCount>0)
+            {
+                for (int i = 0; i < multiselect_list.size(); i++)
+                    docsList.remove(multiselect_list.get(i));
+
+                multiSelectAdapter.notifyDataSetChanged();
+
+                if (mActionMode != null) {
+                    mActionMode.finish();
+                }
+            }
             Toast.makeText(mcontext, FileCount+" file deleted", Toast.LENGTH_SHORT).show();
+            CustomProgressDialog.dismiss();
 
 
         }
@@ -539,11 +608,29 @@ public class DocsActivityRe extends AppCompatActivity implements AlertDialogHelp
         for(int i=0;i<delete_list.size();i++)
         {
             File f=new File(String.valueOf(delete_list.get(i).getFilePath()));
-            if(f.exists())
-                if(f.delete()) {
+            if(f.exists()) {
+                if (f.delete()) {
                     count++;
                     sendBroadcast(f);
                 }
+                //new
+                else {
+                    boolean st = UtilityStorage.isWritableNormalOrSaf(f, mcontext);
+                    System.out.println("" + st);
+                    if (st) {
+                        boolean status = UtilityStorage.deleteWithAccesFramework(mcontext, f);
+                        if (status) {
+                            count++;
+                            Utility.RunMediaScan(mcontext, f);
+                        }
+                    } else {
+                       // UtilityStorage.triggerStorageAccessFramework(mcontext);
+                    }
+
+
+                }
+                //new
+            }
 
         }
         Constants.DELETED_DOCUMENT_FILES=count;
