+++ app/src/main/java/com/example/filehunt/DownloadActivityRe.java
@@ -35,8 +35,10 @@ import com.example.filehunt.Model.Model_Docs;
 import com.example.filehunt.Model.Model_Download;
 import com.example.filehunt.Utils.AlertDialogHelper;
 import com.example.filehunt.Utils.AsynctaskUtility;
+import com.example.filehunt.Utils.CustomProgressDialog;
 import com.example.filehunt.Utils.RecyclerItemClickListener;
 import com.example.filehunt.Utils.Utility;
+import com.example.filehunt.Utils.UtilityStorage;
 
 import java.io.File;
 import java.util.ArrayList;
@@ -78,6 +80,7 @@ public class DownloadActivityRe extends AppCompatActivity implements AlertDialog
         blankIndicator=(ImageView) findViewById(R.id.blankIndicator);
         mcontext=DownloadActivityRe.this;
         instance=this;
+        UtilityStorage.InitilaizePrefs(mcontext);
         Utility.setActivityTitle(mcontext,getResources().getString(R.string.cat_Download));
 
 //           int_position = getIntent().getIntExtra("value", 0);
@@ -115,6 +118,30 @@ public class DownloadActivityRe extends AppCompatActivity implements AlertDialog
 
             }
         }));
+
+
+    }
+
+    @Override
+    protected void onActivityResult(int requestCode, int resultCode, Intent data) {
+        super.onActivityResult(requestCode, resultCode, data);
+
+        if(requestCode==Constants.FILE_DELETE_REQUEST_CODE) {
+            boolean isPersistUriSet = UtilityStorage.setUriForStorage(requestCode, resultCode, data);
+            if (isPersistUriSet && multiselect_list.size() > 0)
+                new DeleteFileTask(multiselect_list).execute();
+        }
+        if(requestCode==Constants.FILE_RENAME_REQUEST_CODE)
+        {
+            boolean isPersistUriSet = UtilityStorage.setUriForStorage(requestCode, resultCode, data);
+            if (isPersistUriSet && multiselect_list.size() > 0)
+            {
+                // call rename function  here in the case of premission granted first  time;
+
+                Utility.renameFile(mcontext,multiselect_list.get(0).getFilePath(),Constants.Global_File_Rename_NewName,4);
+            }
+
+        }
 
 
     }
@@ -341,17 +368,35 @@ public class DownloadActivityRe extends AppCompatActivity implements AlertDialog
         {
             if(multiselect_list.size()>0)
             {
-                new  DeleteFileTask(multiselect_list).execute();
-                for(int i=0;i<multiselect_list.size();i++)
-                    DownloadList.remove(multiselect_list.get(i));
 
-                multiSelectAdapter.notifyDataSetChanged();
+                //
+                File f =new File(multiselect_list.get(0).getFilePath());
 
-                if (mActionMode != null) {
-                    mActionMode.finish();
+                if(UtilityStorage.isWritableNormalOrSaf(f,mcontext)) {
+                    new DeleteFileTask(multiselect_list).execute();
+                }
+                else
+                {
+                    UtilityStorage.guideDialogForLEXA(mcontext,f.getParent(),Constants.FILE_DELETE_REQUEST_CODE);
                 }
 
+
+
+                //
+
+
+                // now this task  is being done  on post execute of delete task
+//                for(int i=0;i<multiselect_list.size();i++)
+//                    DownloadList.remove(multiselect_list.get(i));
+//
+//                multiSelectAdapter.notifyDataSetChanged();
+//
+//                if (mActionMode != null) {
+//                    mActionMode.finish();
+//                }
+
             }
+
         }
         else if(from==2)
         {
@@ -402,6 +447,12 @@ public class DownloadActivityRe extends AppCompatActivity implements AlertDialog
 
     private class DeleteFileTask extends AsyncTask<Void,Void,Integer>
     {
+        @Override
+        protected void onPreExecute() {
+            super.onPreExecute();
+            CustomProgressDialog.show(mcontext,getResources().getString(R.string.deleting_file));
+        }
+
         ArrayList<Model_Download> multiselect_list;
         DeleteFileTask( ArrayList<Model_Download> multiselect_list)
         {
@@ -418,8 +469,20 @@ public class DownloadActivityRe extends AppCompatActivity implements AlertDialog
         protected void onPostExecute(Integer FileCount) {
             super.onPostExecute(FileCount);
 
-            Toast.makeText(mcontext, FileCount+" file deleted", Toast.LENGTH_SHORT).show();
+            if(FileCount>0)
+            {
+                for (int i = 0; i < multiselect_list.size(); i++)
+                    DownloadList.remove(multiselect_list.get(i));
 
+                multiSelectAdapter.notifyDataSetChanged();
+
+                if (mActionMode != null) {
+                    mActionMode.finish();
+                }
+            }
+
+            Toast.makeText(mcontext, FileCount+" file deleted", Toast.LENGTH_SHORT).show();
+            CustomProgressDialog.dismiss();
 
         }
     }
@@ -520,11 +583,29 @@ public class DownloadActivityRe extends AppCompatActivity implements AlertDialog
         for(int i=0;i<delete_list.size();i++)
         {
             File f=new File(String.valueOf(delete_list.get(i).getFilePath()));
-            if(f.exists())
-                if(f.delete()) {
+            if(f.exists()) {
+                if (f.delete()) {
                     count++;
-                   sendBroadcast(f);
+                    sendBroadcast(f);
                 }
+                //new
+                else {
+                    boolean st = UtilityStorage.isWritableNormalOrSaf(f, mcontext);
+                    System.out.println("" + st);
+                    if (st) {
+                        boolean status = UtilityStorage.deleteWithAccesFramework(mcontext, f);
+                        if (status) {
+                            count++;
+                            Utility.RunMediaScan(mcontext, f);
+                        }
+                    } else {
+                        //UtilityStorage.triggerStorageAccessFramework(mcontext);
+                    }
+
+
+                }
+                //new
+            }
 
         }
         Constants.DELETED_DOWNLOAD_FILES=count;
